Setup TS for src of each workspace.
