+++ packages/rbac-http/package.json
@@ -22,8 +22,10 @@
         "axios": "^1.7.7"
     },
     "devDependencies": {
+        "@rollup/plugin-typescript": "^12.1.1",
         "express": "^4.21.1",
         "mocha": "^10.8.2",
-        "rollup": "^4.24.4"
+        "rollup": "^4.24.4",
+        "typescript": "^5.6.3"
     }
 }
