+++ packages/rbac-http/rollup.config.mjs
@@ -1,5 +1,7 @@
+import typescript from '@rollup/plugin-typescript';
+
 export default {
-  input: 'src/index.js',
+  input: 'src/index.ts',
   output: {
     file: 'dist/index.js',
     format: 'umd',
@@ -9,6 +11,7 @@ export default {
       'axios': 'axios'
     }
   },
+  plugins: [typescript()],
   external: [
     '@brainstaff/rbac',
     'axios'
