+++ packages/rbac-in-memory/package.json
@@ -21,7 +21,9 @@
         "@brainstaff/rbac": "^1.0.0-beta.10"
     },
     "devDependencies": {
+        "@rollup/plugin-typescript": "^12.1.1",
         "mocha": "^10.8.2",
-        "rollup": "^4.24.4"
+        "rollup": "^4.24.4",
+        "typescript": "^5.6.3"
     }
 }
