+++ packages/rbac-in-memory/rollup.config.mjs
@@ -1,5 +1,7 @@
+import typescript from '@rollup/plugin-typescript';
+
 export default {
-  input: 'src/index.js',
+  input: 'src/index.ts',
   output: {
     file: 'dist/index.js',
     format: 'umd',
@@ -8,6 +10,7 @@ export default {
       '@brainstaff/rbac': 'rbac'
     }
   },
+  plugins: [typescript()],
   external: [
     '@brainstaff/rbac'
   ]
