+++ packages/rbac-in-memory/src/adapters/RbacInMemoryAssignmentAdapter.ts
@@ -1,4 +1,6 @@
 export default class RbacInMemoryAssignmentAdapter {
+  private rbacAssignments: any;
+
   constructor() {
     this.rbacAssignments = [];
   }
