+++ packages/rbac-in-memory/src/adapters/RbacInMemoryItemAdapter.ts
@@ -1,4 +1,6 @@
 export default class RbacInMemoryItemAdapter {
+  private rbacItems: any;
+
   constructor() {
     this.rbacItems = [];
   }
