+++ packages/rbac-mongodb/package.json
@@ -21,10 +21,12 @@
         "@brainstaff/rbac": "^1.0.0-beta.10"
     },
     "devDependencies": {
+        "@rollup/plugin-typescript": "^12.1.1",
         "chai": "^5.1.2",
         "mocha": "^10.8.2",
         "mongoose": "^8.8.0",
-        "rollup": "^4.24.4"
+        "rollup": "^4.24.4",
+        "typescript": "^5.6.3"
     },
     "peerDependencies": {
         "mongoose": "^8.8.0"
