+++ packages/rbac-mongodb/src/RbacMongodbAdapter.ts
@@ -1,14 +1,19 @@
-import RbacInMemoryAssignmentAdapter from './adapters/RbacInMemoryAssignmentAdapter';
-import RbacInMemoryItemAdapter from './adapters/RbacInMemoryItemAdapter';
-import RbacInMemoryItemChildAdapter from './adapters/RbacInMemoryItemChildAdapter';
-import RbacInMemoryRuleAdapter from './adapters/RbacInMemoryRuleAdapter';
+import RbacMongodbAssignmentAdapter from './adapters/RbacMongodbAssignmentAdapter';
+import RbacMongodbItemAdapter from './adapters/RbacMongodbItemAdapter';
+import RbacMongodbItemChildAdapter from './adapters/RbacMongodbItemChildAdapter';
+import RbacMongodbRuleAdapter from './adapters/RbacMongodbRuleAdapter';
 
-export default class RbacInMemoryAdapter {
-  constructor() {
-    this.assignmentAdapter = new RbacInMemoryAssignmentAdapter();
-    this.itemAdapter = new RbacInMemoryItemAdapter();
-    this.itemChildAdapter = new RbacInMemoryItemChildAdapter();
-    this.ruleAdapter = new RbacInMemoryRuleAdapter();
+export default class RbacMongodbAdapter {
+  private assignmentAdapter: any;
+  private itemAdapter: any;
+  private itemChildAdapter: any;
+  private ruleAdapter: any;
+
+  constructor({}) {
+    this.assignmentAdapter = new RbacMongodbAssignmentAdapter();
+    this.itemAdapter = new RbacMongodbItemAdapter();
+    this.itemChildAdapter = new RbacMongodbItemChildAdapter();
+    this.ruleAdapter = new RbacMongodbRuleAdapter();
   }
 
   /**
@@ -20,10 +25,10 @@ export default class RbacInMemoryAdapter {
   }
 
   async store(rbacHierarchy) {
-    await this.assignmentAdapter.store([...rbacHierarchy.rbacAssignments]);
-    await this.itemAdapter.store([...rbacHierarchy.rbacItems]);
-    await this.itemChildAdapter.store([...rbacHierarchy.rbacItemChildren]);
-    await this.ruleAdapter.store([...rbacHierarchy.rbacRules]);
+    await this.assignmentAdapter.store(rbacHierarchy.rbacAssignments);
+    await this.itemAdapter.store(rbacHierarchy.rbacItems);
+    await this.itemChildAdapter.store(rbacHierarchy.rbacItemChildren);
+    await this.ruleAdapter.store(rbacHierarchy.rbacRules);
   }
 
   async load() {
