+++ packages/rbac-mongodb/src/adapters/RbacMongodbAssignmentAdapter.ts
