+++ packages/rbac-mongodb/src/adapters/RbacMongodbItemAdapter.ts
