+++ packages/rbac-mongodb/src/adapters/RbacMongodbRuleAdapter.ts
