+++ packages/rbac-postgres/package.json
@@ -21,12 +21,14 @@
         "@brainstaff/rbac": "^1.0.0-beta.10"
     },
     "devDependencies": {
+        "@rollup/plugin-typescript": "^12.1.1",
         "chai": "^5.1.2",
         "knex": "^3.1.0",
         "mocha": "^10.8.2",
         "objection": "^3.1.5",
         "pg": "^8.13.1",
-        "rollup": "^4.24.4"
+        "rollup": "^4.24.4",
+        "typescript": "^5.6.3"
     },
     "peerDependencies": {
         "knex": "^0.21.1",
