+++ packages/rbac/rollup.config.mjs
@@ -1,8 +1,11 @@
+import typescript from '@rollup/plugin-typescript';
+
 export default {
-  input: 'src/index.js',
+  input: 'src/index.ts',
   output: {
     file: 'dist/index.js',
     format: 'umd',
     name: 'rbac'
-  }
+  },
+  plugins: [typescript()]
 };
